pyproject.toml
[project]
name = "mmwave-radar-research"
version = "0.1.0"
description = "毫米波雷达底层信号处理与特征提取"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "jupyterlab>=4.5.6",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "open3d>=0.19.0",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pyserial>=3.5",
    "scipy>=1.13.1",
    "torch>=2.8.0",
]
